Moved InjectOnPropertyNameChangedTests into sepearate file. Changed InjectOnPropertyNameChanged to be set directly via property during initialization as opposed to passing in XML config element.
